Enable Railing Buckling (#177)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Simply enables the ability to buckle restrained people to railings.
Useful for when security needs a detainee to hold still (In good faith
:> Reminder you can struggle out IIRC) for a myriad of reasons (Needing
to clear mobs without risking the prisoner, needing to clear a path,
Holding someone still so they will listen to you. **_Other things_**)

## Changelog
:cl:
tweak: Enables can_buckle and buckle_require_restraints for railings. 
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->